PR comments
1. Move `InputGraph` to its module.
